% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivanr.R
\docType{package}
\name{ivanr}
\alias{ivanr}
\alias{ivanr-package}
\title{ivanr: A package for computating the notorious bar statistic.}
\description{
The ivanr package provides three categories of important functions:
ivanr, bar and baz.
}
\section{Foo functions}{

The ivanr functions ...
}

